# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: openai
    description: OpenAI AI chat and image operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerAuth
      tags: [auth]
      summary: Create an account and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterInput"
      responses:
        "201":
          description: Created user with session cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /auth/login:
    post:
      operationId: loginAuth
      tags: [auth]
      summary: Sign in and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /auth/logout:
    post:
      operationId: logoutAuth
      tags: [auth]
      summary: Clear the session cookie
      responses:
        "204":
          description: Logged out

  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get the signed-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List the signed-in user's conversations (newest first)
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          required: false
          description: Id of the last conversation on the previous page
          schema:
            type: number
            minimum: 1
      responses:
        "200":
          description: Page of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/search:
    get:
      operationId: searchOpenaiSources
      tags: [openai]
      summary: Search public sources for grounded answers
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public source results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiSearchResponse"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation (oldest first)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 1000
            default: 200
        - name: cursor
          in: query
          required: false
          description: Id of the last message on the previous page
          schema:
            type: number
            minimum: 1
      responses:
        "200":
          description: Page of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiImageInput"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiImageOutput"

  /openai/stats:
    get:
      operationId: getOpenaiStats
      tags: [openai]
      summary: Get usage stats summary
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password

    AuthRegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - name
        - email
        - password

    AuthUser:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
          format: email
        name:
          type: string
      required:
        - id
        - email
        - name

    OpenaiConversation:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        mode:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - mode
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: number
        conversationId:
          type: number
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
        mode:
          type: string
      required:
        - title

    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
        mode:
          type: string
        context:
          type: string
          description: Optional document or source context used to ground the response
      required:
        - content

    OpenaiSearchSource:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
          format: uri
        snippet:
          type: string
        domain:
          type: string
      required: [title, url, snippet, domain]

    OpenaiSearchResponse:
      type: object
      properties:
        query:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiSearchSource"
      required: [query, sources]

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        mode:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - mode
        - createdAt
        - messages

    OpenaiImageInput:
      type: object
      properties:
        prompt:
          type: string
        size:
          type: string
          enum: ["1024x1024", "1536x1024", "1024x1536"]
      required:
        - prompt

    OpenaiImageOutput:
      type: object
      properties:
        b64_json:
          type: string
      required:
        - b64_json

    OpenaiStats:
      type: object
      properties:
        totalConversations:
          type: number
        totalMessages:
          type: number
        recentConversations:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiConversation"
      required:
        - totalConversations
        - totalMessages
        - recentConversations

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error