# lib/api-zod/src/generated/types/openaiSearchResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OpenaiSearchSource } from './openaiSearchSource';

export interface OpenaiSearchResponse {
  query: string;
  sources: OpenaiSearchSource[];
}