# lib/api-zod/src/generated/types/openaiSearchSource.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiSearchSource {
  title: string;
  url: string;
  snippet: string;
  domain: string;
}